Created simulator module
