+++ CMakeLists.txt
@@ -19,12 +19,12 @@ enable_testing()
 add_executable(
         hello_test
         assember_test.cpp fileworker.cpp fileworker.h syntaxanalyzer.cpp syntaxanalyzer.h int24_t.cpp int24_t.h info.h
-        commands.cpp commands.h codegenerator.cpp codegenerator.h)
+        commands.cpp commands.h codegenerator.cpp codegenerator.h simulator.cpp simulator.h)
 
 add_executable(
         KursovaAK
         main.cpp
-        fileworker.cpp fileworker.h syntaxanalyzer.cpp syntaxanalyzer.h int24_t.cpp int24_t.h info.h commands.cpp commands.h codegenerator.cpp codegenerator.h)
+        fileworker.cpp fileworker.h syntaxanalyzer.cpp syntaxanalyzer.h int24_t.cpp int24_t.h info.h commands.cpp commands.h codegenerator.cpp codegenerator.h simulator.cpp simulator.h)
 
 target_link_libraries(
         hello_test
